requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
requests==2.31.0
aiofiles==23.2.1 
xmltodict==0.13.0
PyYAML==6.0.2
chardet==5.2.0